+++ lib/godwoken_indexer/worker/erc721_updater.ex
@@ -1,48 +1,33 @@
 defmodule GodwokenIndexer.Worker.ERC721Updater do
   use Oban.Worker, queue: :default
 
-  import Ecto.Query, only: [from: 2]
-
-  alias GodwokenExplorer.{Repo, UDT}
-  alias GodwokenExplorer.Token.MetadataRetriever
   import GodwokenRPC.Util, only: [import_timestamps: 0]
   alias GodwokenExplorer.Chain.{Import}
+  alias GodwokenExplorer.Token.MetadataRetriever
 
   @impl Oban.Worker
-
-  def perform(%Oban.Job{}) do
-    do_perform()
-    :ok
-  end
-
-  def do_perform() do
-    datetime = Timex.now() |> Timex.shift(minutes: -5)
-
-    unfetched_udts =
-      from(u in UDT,
-        where:
-          u.type == :native and u.eth_type == :erc721 and
-            (is_nil(u.name) or is_nil(u.symbol)) and u.updated_at < ^datetime,
-        limit: 50,
-        order_by: [desc: u.id]
-      )
-      |> Repo.all()
-
-    unfetched_udts_hashes = unfetched_udts |> Enum.map(& &1.contract_address_hash)
-
-    {:ok, return} = MetadataRetriever.get_functions_of(unfetched_udts_hashes)
+  def perform(%Oban.Job{args: %{chunk_unfetched_udts: chunk_unfetched_udts}}) do
+    chunk_unfetched_udts_hashes = chunk_unfetched_udts |> Enum.map(& &1.contract_address_hash)
+    {:ok, return} = MetadataRetriever.get_functions_of(chunk_unfetched_udts_hashes)
 
     need_update_list =
-      Enum.zip(unfetched_udts, return)
+      Enum.zip(chunk_unfetched_udts, return)
       |> Enum.map(fn {u, r} ->
         if u.contract_address_hash == r.contract_address_hash do
           Map.from_struct(u)
           |> Map.merge(r)
-          |> Map.delete(:__meta__)
         else
           Map.from_struct(u)
-          |> Map.delete(:__meta__)
         end
+        |> Map.take([
+          :id,
+          :name,
+          :totalSupply,
+          :decimals,
+          :symbol,
+          :update_at,
+          :contract_address_hash
+        ])
       end)
 
     Import.insert_changes_list(
@@ -52,5 +37,7 @@ defmodule GodwokenIndexer.Worker.ERC721Updater do
       on_conflict: {:replace, [:name, :symbol, :updated_at]},
       conflict_target: :id
     )
+
+    :ok
   end
 end
